Moving external childresource to resources package
