+++ azure-mgmt-compute/src/main/java/com/microsoft/azure/management/compute/VirtualMachineExtension.java
@@ -1,6 +1,7 @@
 package com.microsoft.azure.management.compute;
 
 import com.microsoft.azure.management.compute.implementation.VirtualMachineExtensionInner;
+import com.microsoft.azure.management.resources.fluentcore.arm.models.ExternalChildResource;
 import com.microsoft.azure.management.resources.fluentcore.model.Attachable;
 import com.microsoft.azure.management.resources.fluentcore.model.Settable;
 import com.microsoft.azure.management.resources.fluentcore.model.Wrapper;
