--- azure-mgmt-compute/src/test/java/com/microsoft/azure/management/compute/childresource/Pullet.java
@@ -1,6 +1,0 @@
-package com.microsoft.azure.management.compute.childresource;
-
-import com.microsoft.azure.management.compute.ExternalChildResource;
-
-interface Pullet extends ExternalChildResource<Pullet> {
-}
