+++ azure-mgmt-resources/src/test/java/com/microsoft/azure/management/resources/childresource/Pullet.java
@@ -1,0 +1,6 @@
+package com.microsoft.azure.management.resources.childresource;
+
+import com.microsoft.azure.management.resources.fluentcore.arm.models.ExternalChildResource;
+
+interface Pullet extends ExternalChildResource<Pullet> {
+}
